android: move opening hours to own fragment
