+++ android/res/layout/place_page_opening_hours_fragment.xml
@@ -3,7 +3,6 @@
   xmlns:android="http://schemas.android.com/apk/res/android"
   xmlns:app="http://schemas.android.com/apk/res-auto"
   xmlns:tools="http://schemas.android.com/tools"
-  android:id="@+id/ll__place_schedule"
   style="@style/PlacePageItemFrame"
   android:tag="schedule">
